refactor: extract progress bar logic into separate module
