+++ src/commands/watch.ts
@@ -1,16 +1,16 @@
-import type { DisplayOptions, ModelType } from '../types.internal.ts';
+import type { DisplayOptions } from '../types.internal.ts';
 import process from 'node:process';
 import Table from 'cli-table3';
 import { define } from 'gunshi';
 import pc from 'picocolors';
 import { getDefaultClaudePath, loadSessionBlockData } from '../data-loader.ts';
 import { log, logger } from '../logger.ts';
+import { calculateModelBreakdown, createProgressBar, displayCostProgressBar } from '../progress-bar.internal.ts';
 import {
 	type BurnRateAnalysis,
 	calculateBurnRate,
 	calculateBurnRateAnalysis,
 	DEFAULT_SESSION_DURATION_HOURS,
-	type LoadedUsageEntry,
 	projectBlockUsage,
 	type SessionBlock,
 	TIME_CONSTANTS,
@@ -30,9 +30,8 @@ type BlockState = {
 };
 
 /**
- * Constants for progress bars and timing
+ * Constants for timing
  */
-const PROGRESS_BAR_WIDTH = 40;
 const BLOCK_DURATION_MINUTES = DEFAULT_SESSION_DURATION_HOURS * 60;
 
 /**
@@ -42,167 +41,6 @@ const TIME_WARNING_THRESHOLDS = {
 	CRITICAL: 30, // Red warning when 30 minutes or less remaining
 	WARNING: 60, // Yellow warning when 1 hour or less remaining
 };
-
-/**
- * Gets the model type from a model name
- */
-function getModelType(model: string): ModelType | null {
-	if (model.includes('opus')) {
-		return 'opus';
-	}
-	if (model.includes('sonnet')) {
-		return 'sonnet';
-	}
-	if (model.includes('haiku')) {
-		return 'haiku';
-	}
-	return null;
-}
-
-/**
- * Calculates model breakdown data from block entries using actual model names
- */
-function calculateModelBreakdown(entries: LoadedUsageEntry[]): {
-	modelTokens: Map<string, number>;
-	modelCosts: Map<string, number>;
-} {
-	const modelTokens = new Map<string, number>();
-	const modelCosts = new Map<string, number>();
-
-	for (const entry of entries) {
-		const modelName = entry.model;
-
-		const currentTokens = modelTokens.get(modelName) ?? 0;
-		const entryTokens = entry.usage.inputTokens + entry.usage.outputTokens;
-		modelTokens.set(modelName, currentTokens + entryTokens);
-
-		const currentCost = modelCosts.get(modelName) ?? 0;
-		const entryCost = entry.costUSD ?? 0;
-		modelCosts.set(modelName, currentCost + entryCost);
-	}
-
-	return { modelTokens, modelCosts };
-}
-
-/**
- * Creates and displays the combined progress bar for models grouped by type
- */
-function displayCostProgressBar(
-	modelCosts: Map<string, number>,
-	maxCost: number,
-	projection: { totalCost: number } | null,
-): void {
-	// Group costs by model type
-	let opusCost = 0;
-	let sonnetCost = 0;
-	let haikuCost = 0;
-
-	for (const [modelName, cost] of modelCosts.entries()) {
-		const modelType = getModelType(modelName);
-		if (modelType != null) {
-			switch (modelType) {
-				case 'opus':
-					opusCost += cost;
-					break;
-				case 'sonnet':
-					sonnetCost += cost;
-					break;
-				case 'haiku':
-					haikuCost += cost;
-					break;
-			}
-		}
-	}
-	const totalUsedCost = opusCost + sonnetCost + haikuCost;
-
-	// Calculate percentages, capping at 1.0 for display purposes
-	const totalPercentage = Math.min(totalUsedCost / maxCost, 1.0);
-	const opusPercentage = totalUsedCost > 0 ? opusCost / totalUsedCost : 0;
-	const sonnetPercentage = totalUsedCost > 0 ? sonnetCost / totalUsedCost : 0;
-	const haikuPercentage = totalUsedCost > 0 ? haikuCost / totalUsedCost : 0;
-
-	// Calculate widths based on the actual usage up to 100%
-	const usedWidth = Math.round(totalPercentage * PROGRESS_BAR_WIDTH);
-
-	// Distribute the used width proportionally among models
-	let opusWidth = 0;
-	let sonnetWidth = 0;
-	let haikuWidth = 0;
-
-	if (usedWidth > 0 && totalUsedCost > 0) {
-		// Calculate proportional widths
-		opusWidth = opusCost > 0 ? Math.max(1, Math.round(opusPercentage * usedWidth)) : 0;
-		sonnetWidth = sonnetCost > 0 ? Math.max(1, Math.round(sonnetPercentage * usedWidth)) : 0;
-		haikuWidth = haikuCost > 0 ? Math.max(1, Math.round(haikuPercentage * usedWidth)) : 0;
-
-		// Adjust if total exceeds usedWidth due to rounding
-		const totalModelWidth = opusWidth + sonnetWidth + haikuWidth;
-		if (totalModelWidth > usedWidth) {
-			// Reduce the largest width
-			if (opusWidth >= sonnetWidth && opusWidth >= haikuWidth && opusWidth > 1) {
-				opusWidth -= (totalModelWidth - usedWidth);
-			}
-			else if (sonnetWidth >= haikuWidth && sonnetWidth > 1) {
-				sonnetWidth -= (totalModelWidth - usedWidth);
-			}
-			else if (haikuWidth > 1) {
-				haikuWidth -= (totalModelWidth - usedWidth);
-			}
-		}
-
-		// Ensure we don't have negative widths
-		opusWidth = Math.max(0, opusWidth);
-		sonnetWidth = Math.max(0, sonnetWidth);
-		haikuWidth = Math.max(0, haikuWidth);
-	}
-
-	const totalModelWidth = opusWidth + sonnetWidth + haikuWidth;
-	const emptyWidth = Math.max(0, PROGRESS_BAR_WIDTH - totalModelWidth);
-
-	// Build the progress bar with colors
-	const opusBar = pc.blue('■'.repeat(opusWidth));
-	const sonnetBar = pc.cyan('■'.repeat(sonnetWidth));
-	const haikuBar = pc.magenta('■'.repeat(haikuWidth));
-	const emptyBar = pc.gray('■'.repeat(emptyWidth));
-
-	const combinedProgress = `[${opusBar}${sonnetBar}${haikuBar}${emptyBar}]`;
-
-	// Calculate percentages and projections based on cost
-	const totalPercentageText = ((totalUsedCost / maxCost) * 100).toFixed(1);
-
-	let projectedPercentageText = '';
-	if (projection != null) {
-		const projectedPercentage = (projection.totalCost / maxCost) * 100;
-		const projectedPercentageStr = projectedPercentage.toFixed(1);
-		if (projectedPercentage > 100) {
-			projectedPercentageText = ` (Est. ${pc.red(`${projectedPercentageStr}%`)})`;
-		}
-		else {
-			projectedPercentageText = ` (Est. ${projectedPercentageStr}%)`;
-		}
-	}
-
-	// Display the combined progress bar
-	log(`Cost Usage:     ${combinedProgress} ${totalPercentageText}%${projectedPercentageText}`);
-
-	// Show legend - only display models that have been used (based on cost)
-	const legendItems = [];
-	if (opusCost > 0) {
-		const opusPercentageDisplay = ((opusCost / maxCost) * 100).toFixed(1);
-		legendItems.push(`${pc.blue('■')} opus ${opusPercentageDisplay}%`);
-	}
-	if (sonnetCost > 0) {
-		const sonnetPercentageDisplay = ((sonnetCost / maxCost) * 100).toFixed(1);
-		legendItems.push(`${pc.cyan('■')} sonnet ${sonnetPercentageDisplay}%`);
-	}
-	if (haikuCost > 0) {
-		const haikuPercentageDisplay = ((haikuCost / maxCost) * 100).toFixed(1);
-		legendItems.push(`${pc.magenta('■')} haiku ${haikuPercentageDisplay}%`);
-	}
-	legendItems.push(`${pc.gray('■')} Unused`);
-	log(`   ${legendItems.join('  ')}`);
-	log('');
-}
 
 /**
  * Displays tokens breakdown table if enabled
@@ -378,40 +216,6 @@ const UPDATE_INTERVALS = {
 	MEDIUM: 15000, // 15 seconds
 	SLOW: 60000, // 60 seconds (1 minute)
 };
-
-/**
- * Creates a progress bar string with the given parameters
- * @param current - Current value
- * @param max - Maximum value
- * @param style - Style configuration for colors
- * @param style.complete - Function to style completed portion
- * @param style.incomplete - Function to style incomplete portion
- * @param style.warning - Optional function to style warning state
- * @param style.warningThreshold - Optional threshold for warning state (0-1)
- * @returns Formatted progress bar string
- */
-function createProgressBar(
-	current: number,
-	max: number,
-	style: {
-		complete: (str: string) => string;
-		incomplete: (str: string) => string;
-		warning?: (str: string) => string;
-		warningThreshold?: number;
-	} = { complete: pc.green, incomplete: pc.red },
-): string {
-	const percentage = Math.min(current / max, 1);
-	const filledWidth = Math.floor(percentage * PROGRESS_BAR_WIDTH);
-	const emptyWidth = PROGRESS_BAR_WIDTH - filledWidth;
-
-	const isWarning = style.warningThreshold != null && percentage >= style.warningThreshold;
-	const completeColor = isWarning && style.warning != null ? style.warning : style.complete;
-
-	const filled = '■'.repeat(filledWidth);
-	const empty = '■'.repeat(emptyWidth);
-
-	return `[${completeColor(filled)}${style.incomplete(empty)}]`;
-}
 
 /**
  * Compares two block states to detect changes
